requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
av==16.0.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
frozenlist==1.8.0
genai-prices==0.0.47
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-genai==1.53.0
googleapis-common-protos==1.72.0
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
livekit==1.0.20
livekit-agents==1.3.6
livekit-api==1.1.0
livekit-blingfire==1.0.0
livekit-plugins-google==1.3.6
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.3.6
livekit-plugins-tavus==1.3.6
livekit-protocol==1.1.1
logfire-api==4.16.0
markdown-it-py==4.0.0
mcp==1.23.1
mdurl==0.1.2
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.5
openai==2.9.0
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
pillow==12.0.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-ai-slim==1.27.0
pydantic-graph==1.27.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.3
sse-starlette==3.0.3
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.20.0
types-protobuf==6.32.1.20251105
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
